+++ lms/src/main/java/com/lawencon/lms/service/impl/LocationsServiceImpl.java
@@ -3,24 +3,28 @@ package com.lawencon.lms.service.impl;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
 import org.springframework.stereotype.Service;
 
 import com.lawencon.lms.constant.EnumCode;
 import com.lawencon.lms.dao.LocationsDao;
+import com.lawencon.lms.dao.PermissionsDao;
+import com.lawencon.lms.dao.PermissionsRolesDao;
+import com.lawencon.lms.dao.RolesDao;
+import com.lawencon.lms.dao.UsersDao;
 import com.lawencon.lms.dto.locations.SaveLocationsResDto;
 import com.lawencon.lms.dto.locations.UpdateLocationsResDto;
 import com.lawencon.lms.model.Companies;
 import com.lawencon.lms.model.Locations;
+import com.lawencon.lms.model.Permissions;
+import com.lawencon.lms.model.PermissionsRoles;
+import com.lawencon.lms.model.Roles;
 import com.lawencon.lms.model.Users;
 import com.lawencon.lms.service.CompaniesService;
 import com.lawencon.lms.service.LocationsService;
-import com.lawencon.lms.service.UsersService;
 
 @Service
 public class LocationsServiceImpl extends BaseServiceLmsImpl implements LocationsService {
-
-	@Autowired
-	private UsersService usersService;
 
 	@Autowired
 	private LocationsDao locationsDao;
@@ -28,19 +32,45 @@ public class LocationsServiceImpl extends BaseServiceLmsImpl implements Location
 	@Autowired
 	private CompaniesService companiesService;
 
+	@Autowired
+	private UsersDao usersDao;
+
+	@Autowired
+	private RolesDao rolesDao;
+
+	@Autowired
+	private PermissionsDao permissionsDao;
+
+	@Autowired
+	private PermissionsRolesDao permissionsRolesDao;
+
 	@Override
 	public List<Locations> findAll() throws Exception {
-		return locationsDao.findAll();
+		String permissionsCode = "PERMSN53";
+		Boolean validation = validationUsers(permissionsCode);
+		if (validation)
+			return locationsDao.findAll();
+		throw new Exception("Access Denied");
 	}
 
 	@Override
 	public Locations findById(String id) throws Exception {
-		return locationsDao.findById(id);
+
+		String permissionsCode = "PERMSN53";
+		Boolean validation = validationUsers(permissionsCode);
+		if (validation)
+			return locationsDao.findById(id);
+		throw new Exception("Access Denied");
 	}
 
 	@Override
 	public Locations findByCode(String code) throws Exception {
-		return locationsDao.findByCode(code);
+
+		String permissionsCode = "PERMSN53";
+		Boolean validation = validationUsers(permissionsCode);
+		if (validation)
+			return locationsDao.findByCode(code);
+		throw new Exception("Access Denied");
 	}
 
 	@Override
@@ -48,28 +78,26 @@ public class LocationsServiceImpl extends BaseServiceLmsImpl implements Location
 		SaveLocationsResDto saveRes = new SaveLocationsResDto();
 
 		try {
-			Companies companies = companiesService.findByCode(locations.getCompanies().getCompaniesCode());
-			locations.setCompanies(companies);
-
-			Users users = usersService.findById(getIdAuth());
-
-			if (users == null) {
-				throw new IllegalAccessException("You need to login first!");
-			}
-
-			else if (!users.getRoles().getRolesName().equals("SUPER-ADMIN") || users.getIsActive() == false) {
-				saveRes.setMessage("only superAdmin can Insert data!");
-				throw new IllegalAccessException("only superAdmin can Insert data!");
+			String permissionsCode = "PERMSN54";
+			Boolean validation = validationUsers(permissionsCode);
+			if (!validation) {
+				throw new Exception("Access Denied");
 			}
 
 			else {
+				Companies companies = companiesService.findByCode(locations.getCompanies().getCompaniesCode());
 
-				if (locations.getLocationsDeploy() == null) {
+				if (companies == null) {
+					throw new Exception("Companies code not found");
+				}
+
+				else if (locations.getLocationsDeploy() == null) {
 					throw new Exception("locationsDeploy required");
 				}
 
 				else {
 					begin();
+					locations.setCompanies(companies);
 					locations.setCreatedBy(getIdAuth());
 					locations.setLocationsCode(generateCode());
 					locations = locationsDao.saveOrUpdate(locations);
@@ -98,19 +126,23 @@ public class LocationsServiceImpl extends BaseServiceLmsImpl implements Location
 			locationsDb.setCompanies(companies);
 			locationsDb.setUpdatedBy(getIdAuth());
 			locationsDb.setLocationsDeploy(locations.getLocationsDeploy());
-			
-			Users users = usersService.findById(getIdAuth());
-			if (!users.getRoles().getRolesName().equals("SUPER-ADMIN") || users.getIsActive() == false) {
-				updateRes.setMessage("only superAdmin can Update data!");
-				throw new IllegalAccessException("only superAdmin can Update data!");
+
+			String permissionsCode = "PERMSN55";
+			Boolean validation = validationUsers(permissionsCode);
+			if (!validation) {
+				throw new Exception("Access Denied");
 			}
 
 			else {
-				begin();
-				locations = locationsDao.saveOrUpdate(locationsDb);
-				commit();
-				updateRes.setVersion(locations.getVersion());
-				updateRes.setMessage("Inserted");
+				if (locationsDb.getLocationsDeploy() == null) {
+					throw new Exception("locationsDeploy required");
+				} else {
+					begin();
+					locations = locationsDao.saveOrUpdate(locationsDb);
+					commit();
+					updateRes.setVersion(locations.getVersion());
+					updateRes.setMessage("Inserted");
+				}
 			}
 
 		} catch (Exception e) {
@@ -122,11 +154,46 @@ public class LocationsServiceImpl extends BaseServiceLmsImpl implements Location
 
 	@Override
 	public Boolean removeById(String id) throws Exception {
-		return locationsDao.removeById(id);
+
+		String permissionsCode = "PERMSN56";
+		Boolean validation = validationUsers(permissionsCode);
+		if (validation) {
+			try {
+				begin();
+				boolean isDeleted = locationsDao.removeById(id);
+				commit();
+				return isDeleted;
+			} catch (Exception e) {
+				e.printStackTrace();
+				rollback();
+				throw new Exception(e);
+			}
+		}
+
+		throw new Exception("Access Denied");
 	}
 
 	public String generateCode() throws Exception {
 		String generatedCode = EnumCode.LOCATIONS.getCode() + (locationsDao.countData() + 1);
 		return generatedCode;
+	}
+
+	public Boolean validationUsers(String permissionsCode) throws Exception {
+		try {
+			Users users = usersDao.findById(getIdAuth());
+			Roles roles = rolesDao.findById(users.getRoles().getId());
+			Permissions permissions = permissionsDao.findByCode(permissionsCode);
+			List<PermissionsRoles> listPermissionsRoles = permissionsRolesDao.findAll();
+			for (int i = 0; i < listPermissionsRoles.size(); i++) {
+				if (listPermissionsRoles.get(i).getPermissions().getId().equals(permissions.getId())) {
+					if (listPermissionsRoles.get(i).getRoles().getId().equals(roles.getId())) {
+						return true;
+					}
+				}
+			}
+			return false;
+		} catch (NotFoundException e) {
+			throw new Exception(e);
+		}
 	}
 }
